requirements.txt
streamlit>=1.32.0
supabase>=2.0.0
fpdf2>=2.7.0
pandas>=2.0.0
plotly>=5.0.0
openpyxl>=3.1.0
qrcode[pil]>=7.4.0
Pillow>=10.0.0
azure-ai-formrecognizer>=3.3.0
azure-core>=1.29.0